+++ tapis-systemslib/src/main/java/edu/utexas/tacc/tapis/systems/service/AuthUtils.java
@@ -246,7 +246,7 @@ public class AuthUtils
     // If a service request and service is in the allowed list then log message and allow.
     if (rUser.isServiceRequest() && SVCLIST_RESOURCETENANT.contains(svcName))
     {
-      log.trace(LibUtils.getMsgAuth("SYSLIB_AUTH_RESOURCETENANT", rUser, systemId, op.name(), resourceTenant));
+      log.info(LibUtils.getMsgAuth("SYSLIB_AUTH_RESOURCETENANT", rUser, systemId, op.name(), resourceTenant));
       return;
     }
     // Log warning and deny authorization
@@ -271,7 +271,7 @@ public class AuthUtils
     if (rUser.isServiceRequest() && SVCLIST_SHAREDAPPCTX.contains(svcName))
     {
       // An allowed service is setting shared context, log message and allow
-      log.trace(LibUtils.getMsgAuth("SYSLIB_AUTH_SHAREDAPPCTX", rUser, systemId, op.name()));
+      log.info(LibUtils.getMsgAuth("SYSLIB_AUTH_SHAREDAPPCTX", rUser, systemId, op.name()));
       return;
     }
     // Log warning and deny authorization
@@ -322,7 +322,7 @@ public class AuthUtils
     }
     // Not authorized, throw an exception
     String msg = LibUtils.getMsgAuth("SYSLIB_PRF_UNAUTH", rUser, name, op.name());
-    log.warn(msg);
+    log.info(msg);
     throw new ForbiddenException(msg);
   }
 
@@ -477,7 +477,7 @@ public class AuthUtils
     if (system == null)
     {
       String msg = LibUtils.getMsgAuth(NOT_FOUND, rUser, systemId);
-      log.warn(msg);
+      log.info(msg);
       throw new NotFoundException(msg);
     }
 
@@ -682,6 +682,7 @@ public class AuthUtils
     if (!rUser.isServiceRequest())
     {
       String msg = LibUtils.getMsgAuth("SYSLIB_UNAUTH", rUser, systemId, op.name());
+      // This is unexpected, so log as warning instead of info
       log.warn(msg);
       throw new ForbiddenException(msg);
     }
@@ -697,6 +698,7 @@ public class AuthUtils
       if (SVCLIST_GETCRED.contains(svcName)) return;
       // Not authorized, throw an exception
       String msg = LibUtils.getMsgAuth("SYSLIB_UNAUTH_GETCRED", rUser, systemId, op.name());
+      // No user or unauthorized service should be doing this, so log as warn instead of info
       log.warn(msg);
       throw new ForbiddenException(msg);
     }
@@ -752,6 +754,7 @@ public class AuthUtils
       case getCred:
         // Only some services allowed to get credentials. Never a user.
         String msg = LibUtils.getMsgAuth("SYSLIB_UNAUTH_GETCRED", rUser, systemId, op.name());
+        // No user should be doing this, so log as warn instead of info
         log.warn(msg);
         throw new ForbiddenException(msg);
     }
@@ -806,7 +809,7 @@ public class AuthUtils
     }
     // Not authorized, throw an exception
     String msg = LibUtils.getMsgAuth("SYSLIB_UNAUTH", rUser, systemId, op.name());
-    log.warn(msg);
+    log.info(msg);
     throw new ForbiddenException(msg);
   }