+++ CHANGELOG.md
@@ -9,10 +9,22 @@ https://tapis.readthedocs.io/en/latest/technical/systems.html
 You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
-## 1.6.4 - 2024-??-??
+## 1.6.5 - 2024-??-??
 
 Incremental improvements.
-Add logging of warning messages when authorization is denied.
+Update service code to match latest System client model.
+
+### New features:
+- None
+
+### Bug fixes:
+- None
+
+---------------------------------------------------------------------------
+## 1.6.4 - 2024-08-05
+
+Incremental improvements.
+Add log messages for when access is denied or resource not found.
 
 ### New features:
 - None